requirements.txt
streamlit
langchain
faiss-cpu
PyPDF2
transformers>=4.30.0
torch>=2.0.0
huggingface-hub
PyMuPDF
sentence-transformers>=2.2.0
langchain-community
pydantic>=2.1
pdf2image
pytesseract
Pillow

